+++ EXPOTECTINA2025/src/main/java/DEV_EXPOTECTINA2025/EXPOTECTINA2025/Repositories/EstadoViajeRepository.java
@@ -3,5 +3,5 @@ package DEV_EXPOTECTINA2025.EXPOTECTINA2025.Repositories;
 import DEV_EXPOTECTINA2025.EXPOTECTINA2025.Entities.EstadoViajeEntities;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadoViajeRepository extends JpaRepository<EstadoViajeEntities, Integer> {
+public interface EstadoViajeRepository extends JpaRepository<EstadoViajeEntities, Long> {
 }
